+++ src/main/java/zenpad/core/Toolbar.java
@@ -4,8 +4,8 @@ import javax.swing.JFrame;
 import javax.swing.JToolBar;
 
 import zenpad.controllers.ToolbarController;
+import zenpad.note.NotePanel;
 import zenpad.runners.CodeRunner;
-import zenpad.ui.NotePanel;
 import zenpad.ui.TabManager;
 import zenpad.views.ToolbarView;
