+++ src/main/java/zenpad/fileopener/FileOpenerController.java
@@ -5,8 +5,8 @@ import javax.swing.event.TreeSelectionEvent;
 import javax.swing.event.TreeSelectionListener;
 import javax.swing.tree.DefaultMutableTreeNode;
 
+import zenpad.note.NotePanel;
 import zenpad.runners.CodeRunner;
-import zenpad.ui.NotePanel;
 import zenpad.ui.TabManager;
 import zenpad.utils.RTextHelper;
