+++ src/main/java/zenpad/fileopener/FileOpenerPanel.java
@@ -2,7 +2,7 @@ package zenpad.fileopener;
 
 import javax.swing.JPanel;
 
-import zenpad.ui.NotePanel;
+import zenpad.note.NotePanel;
 import zenpad.ui.TabManager;
 
 /**
